Move dagger.Factory to dagger.internal.Factory. ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=89607481
