+++ compiler/src/main/java/dagger/internal/codegen/ComponentGenerator.java
@@ -35,9 +35,9 @@ import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import com.google.common.util.concurrent.ListenableFuture;
 import dagger.Component;
-import dagger.Factory;
 import dagger.MapKey;
 import dagger.MembersInjector;
+import dagger.internal.Factory;
 import dagger.internal.InstanceFactory;
 import dagger.internal.MapFactory;
 import dagger.internal.MapProviderFactory;
