+++ compiler/src/main/java/dagger/internal/codegen/FactoryGenerator.java
@@ -21,9 +21,9 @@ import com.google.common.base.Optional;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Lists;
-import dagger.Factory;
 import dagger.MembersInjector;
 import dagger.Provides.Type;
+import dagger.internal.Factory;
 import dagger.internal.codegen.writer.ClassName;
 import dagger.internal.codegen.writer.ClassWriter;
 import dagger.internal.codegen.writer.ConstructorWriter;
