+++ compiler/src/test/java/dagger/internal/codegen/InjectConstructorFactoryGeneratorTest.java
@@ -129,7 +129,7 @@ public final class InjectConstructorFactoryGeneratorTest {
     JavaFileObject expected = JavaFileObjects.forSourceLines("test.GenericClass$$Factory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "import javax.inject.Provider;",
         "",
@@ -173,8 +173,8 @@ public final class InjectConstructorFactoryGeneratorTest {
     JavaFileObject expected = JavaFileObjects.forSourceLines("test.GenericClass$$Factory",
         "package test;",
         "",
-        "import dagger.Factory;",
         "import dagger.MembersInjector;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "",
         "@Generated(\"dagger.internal.codegen.ComponentProcessor\")",
@@ -216,7 +216,7 @@ public final class InjectConstructorFactoryGeneratorTest {
     JavaFileObject expected = JavaFileObjects.forSourceLines("test.GenericClass$$Factory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "",
         "@SuppressWarnings(\"rawtypes\")",
@@ -253,7 +253,7 @@ public final class InjectConstructorFactoryGeneratorTest {
     JavaFileObject expected = JavaFileObjects.forSourceLines("test.GenericClass$$Factory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "import javax.inject.Provider;",
         "",
@@ -300,7 +300,7 @@ public final class InjectConstructorFactoryGeneratorTest {
     JavaFileObject expected = JavaFileObjects.forSourceLines("test.GenericClass$$Factory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import java.util.List;",
         "import javax.annotation.Generated;",
         "import javax.inject.Provider;",
@@ -360,8 +360,8 @@ public final class InjectConstructorFactoryGeneratorTest {
     JavaFileObject expected = JavaFileObjects.forSourceLines("test.GenericClass$$Factory",
         "package test;",
         "",
-        "import dagger.Factory;",
         "import dagger.internal.DoubleCheckLazy;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "import javax.inject.Provider;",
         "",
@@ -621,7 +621,7 @@ public final class InjectConstructorFactoryGeneratorTest {
         "test.InjectConstructor$$Factory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "import javax.inject.Provider;",
         "",
@@ -664,8 +664,8 @@ public final class InjectConstructorFactoryGeneratorTest {
         "test.AllInjections$$Factory",
         "package test;",
         "",
-        "import dagger.Factory;",
         "import dagger.MembersInjector;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "import javax.inject.Provider;",
         "",
@@ -719,8 +719,8 @@ public final class InjectConstructorFactoryGeneratorTest {
         "test.B$$Factory",
         "package test;",
         "",
-        "import dagger.Factory;",
         "import dagger.MembersInjector;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "",
         "@Generated(\"dagger.internal.codegen.ComponentProcessor\")",
@@ -764,7 +764,7 @@ public final class InjectConstructorFactoryGeneratorTest {
         "test.InjectConstructor$$Factory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import java.util.List;",
         "import javax.annotation.Generated;",
         "import javax.inject.Provider;",
@@ -813,7 +813,7 @@ public final class InjectConstructorFactoryGeneratorTest {
         "test.InjectConstructor$$Factory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "import javax.inject.Provider;",
         "",
@@ -864,7 +864,7 @@ public final class InjectConstructorFactoryGeneratorTest {
         "test.InjectConstructor$$Factory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "import javax.inject.Provider;",
         "import other.pkg.Outer;",
@@ -918,7 +918,7 @@ public final class InjectConstructorFactoryGeneratorTest {
         "test.InjectConstructor$$Factory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "import javax.inject.Provider;",
         "import other.pkg.CommonName;",
@@ -968,7 +968,7 @@ public final class InjectConstructorFactoryGeneratorTest {
     JavaFileObject factory = JavaFileObjects.forSourceLines("test.SimpleType$$Factory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "",
         "@Generated(\"dagger.internal.codegen.ComponentProcessor\")",
@@ -1013,7 +1013,7 @@ public final class InjectConstructorFactoryGeneratorTest {
         "test.OuterType$A$$Factory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "import test.OuterType.A;",
         "@Generated(\"dagger.internal.codegen.ComponentProcessor\")",
