+++ compiler/src/test/java/dagger/internal/codegen/ModuleFactoryGeneratorTest.java
@@ -387,7 +387,7 @@ public class ModuleFactoryGeneratorTest {
     JavaFileObject factoryFile = JavaFileObjects.forSourceLines("TestModule$$ProvideStringFactory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "",
         "@Generated(\"dagger.internal.codegen.ComponentProcessor\")",
@@ -433,7 +433,7 @@ public class ModuleFactoryGeneratorTest {
     JavaFileObject factoryFile = JavaFileObjects.forSourceLines("TestModule$$ProvideStringFactory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "",
         "@Generated(\"dagger.internal.codegen.ComponentProcessor\")",
@@ -474,7 +474,7 @@ public class ModuleFactoryGeneratorTest {
     JavaFileObject factoryFile = JavaFileObjects.forSourceLines("TestModule$$ProvideStringFactory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "",
         "@Generated(\"dagger.internal.codegen.ComponentProcessor\")",
@@ -555,8 +555,8 @@ public class ModuleFactoryGeneratorTest {
         "TestModule$$ProvideObjectsFactory",
         "package test;",
         "",
-        "import dagger.Factory;",
         "import dagger.MembersInjector;",
+        "import dagger.internal.Factory;",
         "import java.util.List;",
         "import javax.annotation.Generated;",
         "import javax.inject.Provider;",
@@ -627,7 +627,7 @@ public class ModuleFactoryGeneratorTest {
     JavaFileObject factoryFile = JavaFileObjects.forSourceLines("TestModule$$ProvideStringFactory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import java.util.Collections;",
         "import java.util.Set;",
         "import javax.annotation.Generated;",
@@ -674,7 +674,7 @@ public class ModuleFactoryGeneratorTest {
     JavaFileObject factoryFile = JavaFileObjects.forSourceLines("TestModule$$ProvideStringsFactory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import java.util.Set;",
         "import javax.annotation.Generated;",
         "",
@@ -914,7 +914,7 @@ public class ModuleFactoryGeneratorTest {
         "test.ParentModule$$ProvidesListBFactory",
         "package test;", 
         "", 
-        "import dagger.Factory;", 
+        "import dagger.internal.Factory;",
         "import java.util.List;", 
         "import javax.annotation.Generated;", 
         "import javax.inject.Provider;", 
@@ -951,7 +951,7 @@ public class ModuleFactoryGeneratorTest {
         "test.ChildNumberModule$$ProvideNumberFactory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "",
         "@Generated(\"dagger.internal.codegen.ComponentProcessor\")",
@@ -980,7 +980,7 @@ public class ModuleFactoryGeneratorTest {
         "test.ChildIntegerModule$$ProvideIntegerFactory",
         "package test;",
         "",
-        "import dagger.Factory;",
+        "import dagger.internal.Factory;",
         "import javax.annotation.Generated;",
         "",
         "@Generated(\"dagger.internal.codegen.ComponentProcessor\")",
