+++ core/src/main/java/dagger/internal/Factory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dagger;
+package dagger.internal;
 
 import javax.inject.Inject;
 import javax.inject.Provider;
