+++ core/src/main/java/dagger/internal/MapProviderFactory.java
@@ -15,7 +15,6 @@
  */
 package dagger.internal;
 
-import dagger.Factory;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import javax.inject.Provider;
