+++ core/src/main/java/dagger/internal/SetFactory.java
@@ -15,7 +15,6 @@
  */
 package dagger.internal;
 
-import dagger.Factory;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
