+++ core/src/test/java/dagger/internal/InstanceFactoryTest.java
@@ -15,7 +15,6 @@
  */
 package dagger.internal;
 
-import dagger.Factory;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
